Removing superfluous id property from IndexSetUpdateRequest.
The id property of the IndexSetUpdateRequest DTO is unneeded, as the id
of the index set to update is already given in the URL. It is actually
misleading and requires an additional unneeded check to see if the ids
of the URL and the body are consistent.
Therefore the id is removed from the IndexSetUpdateRequest and the one
from the URL is used only.