Creating Mongo indices in services which do conditional queries.
In this PR, I am trying to create some MongoDB indices for services
doing conditional queries. I also took the chance and renamed the
alarm callback value/service class implementations and extracted field
constants being reused in several places.

This is probably not a complete fix for #2821, but it already helps a
bit.