Adapt grace period to latest changes in alerts
As it is still not possible to mute alerts, we are keeping the grace
period and updating the way it works:

Instead of calculating the grace period using the `triggered_at` time,
we are changing it to use `resolved_at`. That means that the grace period
will prevent to trigger a new alert once that alert is marked as resolved.