Go back in history instead to users page when canceling user form.
The user form component is used in two distinct places:

  * when editing users from the users list
  * when editing the own profile from the top menu

Before this change, when canceling the user form, the user was
redirected to the user list. When missing the required permissions for
this, this fails.

After this change, we are going back one step in the user's history
instead, which goes back to the users page in the first and the place
where the user came from in the second case, which is much more
unsurprising.

Fixes #3075