Keep stream filter after editing stream.
Before this change, whenever a stream was edited and the stream form
modal was closed, the full list of streams was shown again, even if a
filter had been entered. Getting the filtered view again required
pressing the "filter" button.

After this change, whenver the stream form modal is closed, the stream
filter is maintained.

Fixes #2545