Keep calling convention of SearchPage._refreshData consistent. (#3357)
SearchPage._refreshData is called from several places, internally from
the component and when actions of the Decorator dispatcher complete. In
the latter case, the parameter passed to _refreshData is wrong, it's the
decorator id instead of the stream id which scopes the current search.

This change wraps calling _refreshData from the dispatcher so the
parameter is correct.

Fixes #3355.