Automatically unsubscribing from DecoratorsStore in SearchPage.
Before this change, every mount of the SearchPage component included
subscribing to several DecoratorsActions. During an unmount of the
component, no unsubscribing from the actions took place, leaving
dangling calls to a function of an unmounted component behind.

This PR changes the way these actions are subscribed to and uses Reflux
mechanisms to handle unsubscribing automatically.

Fixes #3355.