Unregistering component callbacks from Stream(Rules)Store. (#3378)
* Unregistering component callbacks from Stream(RulesStore).

Before this change, the StreamComponent was registering callbacks to be
notified when the Stream(Rules)Store state changes. Unfortunately, there
was no way to unregister these callbacks, so after unmounting the
component its callbacks were still called.

After this change, there is now a way to unregister callbacks which is
consumed in the componentWillUnmount function of the StreamComponent.

Fixes #3374.

* Explicitly adding lodash as dependency of the web interface.

It was missing before, although used in at least one line of code. The
reason why it worked before is because it was drawn it transitively
through other libraries depending on it.

* Unregistering from stores upon unmount in StreamRulesEditor as well.