Use consistent case in old message count conditions
Since 2.2.0, we are storing the message count `threshold_type` parameter in
uppercase, just like we do with other types in field value conditions.
Previous Graylog versions were storing `threshold_type` in lowercase,
same case that were use in parameters when instantiating new
MessageCountAlertCondition objects. This resulted in case inconsistencies
in objects from that class, which were propagated to the API and
persistance layer.

This commit converts `threshold_type` to uppercase, so we use that case
when instantiating old conditions.

Fixes #3451