Checking if LoginPage is mounted before updating loading indicator. (#3500)
This change prevents a warning showing up in the console every login.
The loading indicator flag is updated from the promise when the
component was already unmounted. This is prevented by checking
`isMounted()` before calling `setState` from the promise handler.
Although this is considered to be an antipattern (see
https://facebook.github.io/react/blog/2015/12/16/ismounted-antipattern.html)
there is currently no real alternative.