Give an option to repeat alert notifications (#3536)
* Add option to repeat notifications

Refs #3511

* Resend notifications when user sets the flag

When user sets the `repeat_notifications` flag in the alert condition,
resend the notifications regardless of the alert state. The alert will
continue to have a state anyway, and all sent notifications will be
visible in the web interface on the alerts detail page.

Refs #3511

* Add repetition setting to alert summaries

* Add repetition setting to alert timeline

* Add date to notification histories

- Sort histories by creation datetime
- Add the time were the notification was sent, to help seeing the order
and also that the attempts sending the notification are different

* Do not include ms in history datetime

* Add license header

* Rename variable

* Fix import

* Use moment to sort notifications by date

(cherry picked from commit 003bea90a4e67c15fca557c96fcd2665c23f90cc)