Restore custom bootstrap theme after using stock version.
Today I noticed that the colors in the web interface are "slightly" off,
after removing static assets and using the npm module version of
bootstrap. I found out that we were using a customized version of
bootstrap, generated by the bootstrap customizer
(http://getbootstrap.com/customize/). I also found out, that the
`bootstrap-config.json` file is from a completely different
customization and looks even weirder.

With this change, we are now switching over to the less-version of the
bootstrap stylesheets and using the less-loader to override customized
variables extracted from the
[customization run](https://github.com/Graylog2/graylog2-server/blob/2.2/graylog2-web-interface/public/stylesheets/bootstrap.min.css#L8-L9).