Fix cloning alert conditions with identical id when cloning stream.
Before this change, any alert conditions associated with a stream were
cloned identically (including using the same id) when cloning a stream.
This was leading to a bug, where the alert condition was shown as being
associated with the source of the clone process instead of the target.

With this change, the alert condition is recreated during cloning,
including generating a new id for it.

Fixes #3608.