Update to react-bootstrap 0.30 (#3598)
* Bump react-bootstrap dependency to 0.29

This allows us to work on changes without breaking everything

* Fix changes in react-bootstrap onSelect callbacks

https://github.com/react-bootstrap/react-bootstrap/blob/master/CHANGELOG.md#v0290

* Remove unnecessary navItem props

* Replace ButtonInput elements with Button

* Introduce Input adapter component

There have been many API changes in react-bootstrap that impeeded us to
update. As changing all inputs at once would require lots of effort,
here I introduce an adapter component that will transform the old API
into the new one.

I also added a deprecation warning when using this component, so copy
and paste will be a bit more discouraged.

The new form API can be found in:
https://react-bootstrap.github.io/components.html#forms

* Fix error when clicking on open indices modal

Since react-bootstrap 0.29, the event handlers used for `onClick` and
`onSelect` expect the event argument to be passed in a different position.

This change will check if any of the two first arguments is an event,
and stop its propatation if that's the case.

* Also accept numbers as Input values

* Add support for horizontal forms and validations

* Only display Input warning when the file is imported

Otherwise the deprecation warnings are too noisy in the browser console.

* Add getChecked function to Input

* Use onChange instead of onClick for checkbox

* Replace deprecated FormControls.Static

* Add more types to allowed InputWrapper children

* Add missing IDs to Tabs elements

* Add a bit more room for data filters with longer label

* Use Input wrapper in typeahead components

* Update to react-bootstrap 0.30.x

* Replace deprecated PageItem with Pager.Item

* Use inputRef to get form input reference

Avoids using the discouraged `ReactDOM.findDOMNode()` function.

* Remove unused import

* Retention/rotation inputs should not be standalone

* Add support for addon after in Input adaptor

* Remove deprecation warning for Input adaptor

Using the new react-bootstrap input API feels cumbersome, so we should
come up with some API that works for us and that we can use in most of
our cases.