Restore removal of role permissions upon roles update.
In #2529, while fixing #2516, persisting all permissions of roles
assigned to a user was reintroduced (fixed in
35c9325525ad18284411ebdbd50d2c62eaed6732 before). This happened due to
the introduction of a temporary variable that was not used effctively.

After this change, LDAP permissions are not modified blindly but the
permission set of a user still stays denormalized when modified and
saved.