Better timerange for Show Received Messages button (#3725)
The old behaviour was to link to a search that only looked at the last 8 hours of data. I've come across a whole bunch of users who had messages with wrong timestamps and would only find them when searching in "All messages". This change makes the button link to a search over "All messages" instead, so users with timestamps in the past can find the messages and identify the issue.