Fixing `Indices#exists` and `Indices#aliasExists`.
Before this change, the checks in both methods were too simple:

  * `exists` checked if a HEAD on `/<indexname>` succeeds. This returns
`true` if the supplied name is an alias (which the method is clearly
supposed not to do according to
https://github.com/Graylog2/graylog2-server/blame/master/graylog2-server/src/main/java/org/graylog2/periodical/IndexRotationThread.java#L128.
Instead we are now getting the settings for the index and check if it
actually is an alias or not.

  * `aliasExists` checked if a HEAD on `/_all/_alias/<indexname>`
succeeds. Unfortunately, this also returns true if the name is an index
and not an alias. Instead we are now doing the same call as for `exists`
and reversing the alias condition.