Fix DateHistogram-related functionality in Searches class
Searches.DateHistogramInterval was using java.time.Period to define its intervals and tried to
convert them to milliseconds in the Searches.DateHistogramInterval#getMillis() method which
triggered an exception ("Cannot convert to Seconds as this period contains years and years
vary in length") when using a period with variable length (such as year, quarter, or year).

This change changes the DateHistogram-related methods in Searches to use the equivalent interval
from the org.elasticsearch.search.aggregations.bucket.histogram.DateHistogramInterval class to
perform aggregations.