Get ES version from cluster, return IndexMapping according to that.
Before this change, users need to explicitly configure the Elasticsearch
major version in their config file. After this change, it is retrieved
upon bindings initialization time by querying the root resource of the
Elasticsearch cluster, extracting the retrieved version string and
configuring the index mapping according to that.

Obviously it queries only a single server in the cluster, assuming that
at least the major version of all nodes in the cluster match. But this
is the same for the explicit configuration setting.