Smaller UI and UX changes (#186)
* Unify navigation across all pages

The navigation that was changing on pages was immensely confusing me. Like already done for other parts of the web interface that are following a similar pattern, this change addresses this by always having the same navigation items and highlighting the currently active one.

* Remove now duplicate button to simulator

This helps a little with the large amount of buttons on the page.

* Fix typo

* Warn when a pipeline isn't connected to streams

Show a prominent warning on the pipeline details page in case the pipeline is not connected to any streams.

* Improve "No stream connection" warning

Include a note that you do not have to connect a pipeline to a stream if you intend to use it for decorators only. Also make clear that pipelines that are connected to a stream will process incoming messages.