Don't check ES cluster health when flushing messages (#3927)
* Don't check ES cluster health when flushing messages

Checking the ES cluster health before flushing messages was important
while using the embedded Elasticsearch client node because it would block
processing until the cluster is available and healthy (YELLOW or GREEN).

After migrating to an HTTP-based Elasticsearch client, this isn't necessary
anymore. The client will simply fail to index the messages.

Additionally, this changeset only marks those messages as committed, which
could be successfully indexed. Before this change, all messages of a batch
were marked as committed and removed from the journal, whether they've been
indexed or not.

* Only committing highest journal offset.