Prevent ClassCastException with invalid timestamp in clone_message() (#192)
* Prevent ClassCastException with invalid timestamp in clone_message()

If the "timestamp" field of a message is invalid, `clone_message()` will fail because
`Message#getTimestamp()` tries to cast this field to a `DateTime` object.

With the changes in this commit, the type of the "timestamp" field will be checked
before accessing it and if it is invalid (i. e. not a `DateTime` object), the current
date and time will be used and the original content will be stored in the "_original_timestamp"
field of the new message.

Fixes Graylog2/graylog2-server#3880

* Convert original (invalid) timestamp to string

* Log warning if message has invalid timestamp

* Rename field to "gl2_original_timestamp"