Make Elasticsearch tests to real integration tests
Tests using Elasticsearch were previously relying on the NoSQLUnit Elasticsearch adapter
to create an embedded Elasticsearch 2.x instance.

Unfortunately embedded Elasticsearch nodes are unsupported since Elasticsearch 5.0.0 so
that these tests now have to rely on an externally spawned Elasticsearch instance.

In this case, we're utilizing the elasticsearch-maven-plugin which spawns an Elasticsearch
node during Maven's "verify" lifecycle phase ("integration-test" to be precise) which can
be used by Graylog's integration tests (ending in "*IT" instead of "*Test").