Adding privileges to user for created dashboard if not admin.
Before this change, a user with sufficient privileges to create a
dashboard could create one which is immediately inaccessible to the user
because the user is missing read/edit privileges to it. This change
checks for the required privileges and adds them to this specific
dashboard if necessary.

(cherry picked from commit 0e41dc77d25917f6290ad8140e0baf6f09093a15)