Replace "deprecated" Counts API with Multi Search with size=0
The Elasticsearch Count API doesn't allow setting the indices in the HTTP request body
and suffers from the issue that users with a large number of indices run into the default
HTTP header size limit of Elasticsearch (4 KB).

This change set replaces the (kind of) deprecated Counts API with a regular search with
size=0 issued via the Elasticsearch Multi Search API which allows specifiying the list
of indices in the request body.

Fixes #4136
Refs elastic/elasticsearch#13928