Remove superfluous config from stacked charts
Stacked charts contained some configuration options that were redundant
and/or incorrect. Namely the timerange information was twice in the
configuration object, and the query appeared in the main configuration
object, even if the query is independent for each one of the time series
in the chart.

This commit fixes the issue by filtering out those redundant fields.

Fixes #4150