Converting DashboardsStore to reflux store.
When adding dashboard permissions to non-admin users dynamically on
dashboard creation, the current user needs to be reloaded. With the old
DashboardsStore this lead to timing issue/race conditions that prevented
new dashboards to show up instantly in most cases.

This change is now converting the DashboardsStore to a Reflux store
similar3 to most other stores, which is also refreshing its state when
the CurrentUser store's state changes and triggers a reload of the
current user upon dashboard creation.