Disabling ESLint indent rule for JSX elements.
Due to conflicting views between ESLint and us on how JSX elements
should be indented, the `indent` rule is now disabled for JSX elements.

It will be reenabled when there is the `first` option for JSX prop
indentation. Otherwise a different, JSX-specific indentation rule
supporting this will be put in place instead.