Deprecate `onValueChange` prop on Select component
The purpose of that prop was to offer some internal API we could rely
on, but there were many accidental usages of `onChange`, which was going
directly to react-select.

In this commit we are deprecating `onValueChange` and start using `onChange`
instead. This allows us to keep the old API, fix code that was directly
using react-select's callback, and avoid accidental pass of props to the
encapsulated component.

It is still possible to get a direct callback from react-select by using
the `onReactSelectChange` prop.