Migrate stream matching from MessageFilter to MessageProcessor
Instead of matching stream rules as part of the legacy message filter chain,
this change set moves it into a message processor which can be independently
configured and moved inside the list of message processors as the users require.

For example it wasn't possible to use the GeoIP processor for retrieving the
geo location of a client and sort the message into a country-specific stream
(which requires the country code) without running the processing pipeline last
which also affects other messages (which might require a different order).