Remove navigation warnings (#4192)
* Fix react-unknown-prop on brand link

Filter out `active` prop on logo link.

* Remove eventKey from Navbar.Collapse

This is not needed and triggers a warning, as it is passed down to the
div doing the collapse.

* Use consistent indentation

* Fix another react-unknown-prop warning

Replace hack to remove active MenuItem state with a custom component.
This makes more clear what we want to achieve and removes the warning
from the console.

* Use react-bootstrap's Badge in navigation bar

Replacing the custom `span` tags with `Badge` also removes a warning
from the web console.

We also remove the inline styles with a Badge.css file, which allows us
to apply different appearances to badges. This is not supported by
react-bootstrap or bootstrap out of the box.