Updating react-select to 1.0.0-rc (#4189)
* Update react-select css path

* Adapt onChange callback to changes in react-select

Before react-select passed the object's value, now it passes the whole
object.

* Filter options passed to react-select

Do not pass options not being accepted by the ReactSelect component,
letting the prop check to itself.

* Deprecate `onValueChange` prop on Select component

The purpose of that prop was to offer some internal API we could rely
on, but there were many accidental usages of `onChange`, which was going
directly to react-select.

In this commit we are deprecating `onValueChange` and start using `onChange`
instead. This allows us to keep the old API, fix code that was directly
using react-select's callback, and avoid accidental pass of props to the
encapsulated component.

It is still possible to get a direct callback from react-select by using
the `onReactSelectChange` prop.

* Also filter `value` prop from react-select

* Use function ref in Select component

* Fix select sm styling

* Fix case when user clears select input

* Update react select to 1.0.0-rc.10

* Adapt Select onChange callback on multi selects

This allows us to encapsulate all related react-select changes in the
Select component.

* Migrate Select's onValueChange props to onChange

* Add now supported autoFocus and required props

* Replace deprecated string ref

* SelectableList should not be clearable

Selected items go into a different component, so there's no need to
clear the select input.

* Support creating options from Select component

This was previously possible by using the `allowCreate` prop, and it
suffered many changes in react-select:

- Now we can get that functionality by using react-select's `Creatable`
component

- In order to display the newly added options, react-select requires
that created values have the same format as options. This requires some
formatting in our side, as our Select component returns and accepts
string values

* Pass string value to Select component

(cherry picked from commit 3cfb581c9b1d6b9609565b45e379403118226f26)