Unify action components of all field analyzers
This change unifies the action components (buttons and dropdowns) of all field analyzer widgets we ship. Those components are found in the top right corner of each widget.

* The "Dismiss Widget" button is now always a close icon and on the far right of all actions. Before it was sometimes directly accessible and sometimes hidden behind a "Customize" drop-down. The change to a close _icon_ makes it stand out more and also saves real-estate. A margin to the left helps identifying it quickly and avoids accidental pressing of it.
* The "Add to dashboard" button has the same and less intrusive button across all widgets now.
* The new Quick Values configuration dialog is accessible directly without going through the "Customize" drop-down. In fact, the "Customize" drop-down only had "Customize" as a menu item after "Dismiss" was moved to the new close widget icon.

Note that there will be a separate commit and PR for the map widget, which is in a separate repository. These changes should be merged one after the other to keep UI consistency.