Allow normal users to create dashboards (#4155)
* Adding privileges to user for created dashboard if not admin.

Before this change, a user with sufficient privileges to create a
dashboard could create one which is immediately inaccessible to the user
because the user is missing read/edit privileges to it. This change
checks for the required privileges and adds them to this specific
dashboard if necessary.

* Converting DashboardsStore to reflux store.

When adding dashboard permissions to non-admin users dynamically on
dashboard creation, the current user needs to be reloaded. With the old
DashboardsStore this lead to timing issue/race conditions that prevented
new dashboards to show up instantly in most cases.

This change is now converting the DashboardsStore to a Reflux store
similar3 to most other stores, which is also refreshing its state when
the CurrentUser store's state changes and triggers a reload of the
current user upon dashboard creation.

* Removing store hydration call, retrieving current user is done async.

The call to `CurrentUserStore.get()` could return `undefined` in case
the `CurrentUserStore` has not finished updating yet. Retrieving the
current user's permission set is done anyway, when the
`CurrentUserStore` is updated through the `currentUserUpdated` function.
Until then, an empty permission set is assumed.

* Fix permissions check in PermissionsMixin

According to https://shiro.apache.org/permissions.html, the permission "foo:bar:*" is identical to "foo:bar" which wasn't previously taken into account.