Show stats of all index sets without pagination (#4211)
Add information and stats about all indices on the "Index Sets" page without accounting for pagination. The reason for this is, that finding out the total size of all indices can be really annoying if you have a lot of index sets.

Closes #4204

(cherry picked from commit 9f60751309885627070fcffbccfe7ebfd575b812)