Implement QuickValues over time widget (#4244)
* Extract histogram boundary parsing into tools

* First stab at terms histogram API endpoints

* Add /terms-histogram endpoint for absolute and keyword searches

* Make interval optional and build a sane default based on the timerange

This uses the same logic we use in the web interface to construct an
interval.

* Initial QuickValues over time visualization

* Make sure the chart container is using the complete width

* Add link to go back from histogram to the overview

* Disable animations

* Rename StackedChartVisualization to QuickValuesHistogramVisualization

This is nowhere near a generic component at the moment.

* Move some code from SearchResource into a helper so we can reuse it

* Move some code from QuickvaluesWidgetStrategy into base class

We will reuse this soon.

* Add QUICKVALUES_HISTOGRAM widget strategy

* Make it possible to add the quickvalues histogram to dashboards

* Move quick values config out of QueryConfiguration

Adding this in #4205 was a mistake. The fields need to be in a separate
file.

* Only show data table limit if not qv histogram

* Remove unused "show_chart_legend"

* Restrict the chart legend to max 5 entries

Otherwise we would overwrite the x axis label.

* Improve QV histogram rendering based on GraphVisualization

* Handle resizing issues

* Fix chart size issue where the chart gets smaller every new render

We have been passing this.CHART_MARGINS into the chart and dc.js was
modifying the values. So on every new chart render the chart got
smaller.

Use an immutable value for this.CHART_MARGINS.

* Account for y label in margin

* Fix terms/termsHistogram bug with stacked fields

Wrap terms query with stacked fields in an additional filter to ensure
that the main and stacked fields actually exist in the message. Otherwise
we would get "null" values in the result.

* Cleanup QuickValuesHistogramVisualization

Removing unused variables and simplify some code.

* Cleanup QuickValues configuration forms

* Fix bucket sorting in quickvalue histogram

(cherry picked from commit d5fd35dd3330bb87174b149ba2a9d54b36b7bb2b)