updated default/example graylog.conf
I realized after questions in the community and training sessions that ppl. tend to change the elasticsearch setting in the configuration file and ask why they are not changed after a restart of graylog. 

My intention here is that ppl can actually read where to find the setting in the web interface. It might be technical not 100% correct (I did not checked that) but it will point the ppl to the point where they can change them in a running Graylog environment.