Improvements in unified analyzer menus (#4280)
* Fix styling of unified analyzer buttons

Fixes #4230

* Open analyzer menus to the left

Dropdown menus on the unified analyzer menus open currently to the right, but
they are very close to the edge of the window and will be cropped. This
commit changes that by opening menus to the left.

(cherry picked from commit f09054acf07df4e8c4991209b7513b73c50800a8)