3.0 changes (#4277)
* Add prop to set row height in ReactGridContainer

* Add extra flag to resize in ReactGridContainer

Now it is possible to restrict widget resize, even if ReactGridContainer is
not locked.

* Support callback when visualization is rendered

Most visualizations use a combination of React and a visualization
library to draw charts. This makes it hard to know when the
visualization is fully rendered into the DOM from outer components,
which is needed in order to calculate the visualization size, among
others.

To solve this problem, this commit adds an optional `onRenderComplete`
callback that consumers of the visualization components may provide
to be informed when the visualization is rendered.

The new `RenderCompletionObserver` component abstracts most of the work
needed in order to get a call when the visualization is rendered for the
first time.

* Option to disable interactivity in visualizations

Sometimes visualizations may not be rendered for a user to interact
with them. Add an option that allow us to:

- Disable tooltip rendering
- Disable animations and transitions

* Use function references in visualizations

Both `ReactDOM.findDOMNode()` and string references are deprecated or
discouraged to use. Remove their usages in visualization components.

* Use svg to draw numeric visualization

* Help batik to render numeric visualization svg

Change way of style image and move some styles inline.

* Let consumers of ReactGrid customize column layout

* Reduce padding after numeric viz

* Print colour legend on quick values table

The legend is disabled by bootstrap on media print, but we want to keep
it when printing or generating a pdf.

* Add prop to limit visualization's height

This let us decide whether the content should scroll or not.

* Extract props in render method for clarity

* Add widget id to numeric visualization container

* Add print styles to quick values visualization

* Fix NumericVisualization size

The width and height svg attributes are used to set the coordinate
system. Instead, we should use CSS properties to set the visualization
height and width.

* Use * as default query in stacked chart legend

* Add prop to disable react-grid-layout animations

* Pass dashboard lock status to widgets

Some widgets need to be locked when being dragged. We now pass the
dashboard lock state down.

* Add SortUtils to encapsulate sort functions

For now:
- Sort asc/desc by date
- Sort asc/desc natural ignoring case

* Fix chart size in FF/Chrome

Using a div seems to help making both browser draw widgets in a similar
size.

* Fix numeric visualization styling

Something got broken while manually editing some merge conflicts.