Replace webpack-dev-server with custom express
Many of us (if not everyone) has problems running webpack-dev-server
with many plugins, as the server keeps disconnecting from the web
browser.

In order to solve that, but keeping the setup easy to setup and start,
this commit adds a custom express server, using webpack-dev-middleware
and webpack-hot-middleware to achieve something similar to what we had
before.