Switch to node 8 / yarn (#4361)
* Fix jquery types dependency

We need version 2 instead of 3.

* Ensure we don't leak files in graylog-web-plugin

Use `files` attributes to whitelist files that should be in node
package, instead of blacklisting in .npmignore. This helps avoiding any
leaks in dotfiles or uploading any unwanted file.

* Switch graylog-web-plugin package.json to yarn

The package can now be used with yarn v1.3, and scripts use yarn as
default. Anyway, the package will still work with npm v4.

* Add jquery package resolutions

This ensures only one copy of `jquery` and `@types/jquery` are installed
when using yarn.

* Replace npm with yarn

* Add updated vendor-modules

* Do not exclude yarn.lock file