Do not clear editor textarea on undo
Reset undo history when the editor is fully loaded to avoid that users
clear textare performing an undo command.

The problem behind this is that the editor is created empty and ReactAce
does a `setValue` to set the given value for the editor. This is
registered as a change by ace, so when pressing undo in an apparently
not-changed input, the input is reset.

Fixes Graylog2/graylog-plugin-pipeline-processor#224