Change state shape in sessions duck
Split reducers, so each one of them only takes care of one state item,
and combine them afterwards to export a reducer that can still handle
all state associated to the session.

Notice that, as we use selectors now, none of the components needed to
be changed once the state shape changed.