Add component generator connecting ES6 class components to stores.
This generator allows ES6 class components to be connected to stores,
although the current version of Reflux allows this only using a mixin,
which is not usable for ES6 classes.

The change provides a `connect()` function which takes the supplied
component class and wraps it in a generated component which connects
itself to the stores defined in the provided mapping. The state supplied
by the stores is then passed to the component as props.

It is also possible to pass a list of actions which are executed by the
generated component when mounted. This allows to call store actions
populating its state.

The last parameter defines if the generated component should wait for
all states to provide a non-undefined state, rendering a spinner if this
has not happened yet. This behavior is disabled per default, unless at
least one initial action is passed or the `waitForStores` parameter is
defined to be true.

Take note that if the generated component waits for the stores' states,
all required actions to populate the stores' states should be executed
outside of the `componentDidMount` function of the supplied component,
because the component is not mounted before all stores' states are
present.