Change Permission Buttons from Toggle to Set and Clear (#4467)
After selecting two streams (or dashboards) in the Edit Role
view, two buttons appeared in the header of the table.
"Toggle read permissions" and "Toggle edit permissions"
which would invert the permissions of a selected streams.

The user would rather set or clear the permissions of the
selected streams.

So I changed the label of the buttons from "Toggle" to
"Set" and "Clear" depending on which state the streams
permissions are. If all permissions are "Allow" the button
will read "Clear" and remove the permissions. If one or
more will have no "Allow" permission, the button will read
"Set" and will set the permission to "Allow" for all selected
streams.

Also add test for PermissionSelector Component and did
some small ESLint fixes.

Fixes #3210