Change MobX data stores organization
The `RootStore` used until now had some shortcomings:
- Stores having dependencies on other stores needed to get to them
through `RootStore`, but it could be possible that the other dependency
was still not fully initialized
- The `SessionStore` has a dependency on `FetchProvider`, and
`FetchProvider` on `SessionStore`. Importing `RootStore` from
`FetchProvider` did not properly work, as the `RootStore` instance was
not fully initialized and could not be imported
- Components needed to know how to get to a certain store in their
`inject` methods. This logic should not belong to the component

As we already solved many of those problems, we are sticking to our
`CombinedProvider`/`StoreProvider` to handle our MobX stores as well.
The main difference of this is that each MobX store has to initialise
itself, and every JS file needs to inject the stores it needs by using
`StoreProvider`, just like we did until now.