Fixing relative time range used by RecentMessageLoader.
The RecentMessageLoader component is trying to limit the range of the
search request it used to get the last received message of an input to
the last hour. For this, it uses the range parameter. This is the
correct parameter name for the backend, but the web interface uses a
small conversion helper that expects the parameter to be named
relative for relative time ranges, according to the naming in the old
web interface.

This fix is changing the conversion helper to actually also take the
`range` parameter into account and use it if present, otherwise it uses
the `relative` parameter.

Fixes #4510.