Fixing relative time range used by RecentMessageLoader. (#4513)
The RecentMessageLoader component is trying to limit the range of the
search request it used to get the last received message of an input to
the last hour. For this, it uses the `range` parameter. This is the
correct parameter name for the backend, but the web interface uses a
small conversion helper that expects the parameter to be named
`relative` for relative time ranges, according to the naming in the old
web interface.

To make the fix as small as possible for a point release, this change
just uses the `relative` parameter. For `master` and subsequent releases
we should either switch over to the `range` parameter or use it in case
the `relative` parameter is not present. This way we do not discourage
consumers of the `UniversalSearchStore` to use the actually correct
parameter naming.

Fixes #4510.