Properly handle timestamp field in stacked quickvalues (#4516)
* pass all stacked fields into QuickValuesVisualization

with that information we can properly format the timestamp even if it isn't the first field

still has issues with timezone conversions

* Pass timestamp to format directly

Stacked fields have a different object structure as regular fields, so
we can't rely on an object key being there.

* Do not convert timestamp to UTC

`DateTime` will convert the value to the user timezone, which is what we
want to display.

* include stacked fields in quick values title

use correct default prop name

* create stacked fields array on the fly if the caller doesn't give us the fields

this is the case for widgets, because the container doesn't know what to pass

Fixes #4509

(cherry picked from commit 272604abce7a13f9facf0559c35f0aed0c6d8c56)