Prevent ldap server password leak
The ldap configuration page was setting the input value of the
password field to the actuall password. This way a attacker
could read the ldap password, if he gained admin access to graylog.

This change prevents it by setting the input value of the password
with a placeholder in form of 13 asterisks.

When the gui sends the form field to the server, the server
compares the value with the placeholder and only update the
password if it differs with the placeholder.