Review fixes
Instead of using the magic string "**********" we send the state
of the password in the GET action. So we tell the GUI if a
password is set or not. On the client side we use that information
to set the placeholder accordingly.

But on the server side we accept the system_password_set flag
so we do not have to change the data before sending to
the server.